creating components for the check and delete icons
