+++ script.js
@@ -1,4 +1,5 @@
-( ()=> {
+import checkComplete from "./components/checkComplete.js";
+import deleteIcon from "./components/deleteIcon.js";
 
 const btn = document.querySelector('[data-form-btn]');
 
@@ -26,32 +27,3 @@ const createTask = (evento) => {
 
 btn.addEventListener('click', createTask);
 
-const checkComplete = () => {
-    const i = document.createElement("i");
-    i.classList.add("far", "fa-check-square", "icon");
-    i.addEventListener("click", completeTask)
-    
-    return i;
-};
-
-const completeTask = (event) => {
-    const element = event.target;
-    element.classList.toggle("fas");
-    element.classList.toggle("completeIcon");
-    element.classList.toggle("far");
-};
-
-const deleteIcon = () => {
-    const i = document.createElement("i");
-    i.classList.add("fas", "fa-trash-alt", "trashIcon", "icon");
-    i.addEventListener("click", deleteTask)
-
-    return i
-};
-
-const deleteTask = (event) => {
-    const parent = event.target.parentElement;
-    parent.remove();
-};
-
-})()
